site-recommendation-model/requirements.txt
# Requirements for Street View Image Collection
requests>=2.28.0
numpy>=1.21.0
pillow>=9.0.0
pandas>=1.5.0
matplotlib>=3.5.0
googlemaps>=4.6.0
folium>=0.12.0
geopandas>=0.11.0
python-dotenv>=1.0.0
